+++ lua/plugins/conform.lua
@@ -10,7 +10,7 @@ return {
 					return
 				end
 				return {
-					timeout_ms = 500,
+					timeout_ms = 1500,
 					lsp_format = "fallback",
 				}
 			end,